Fixed the apt_install and related functions, which dmitry broke by making them take only one parameter vs. a list. Dmitry had called them like this:    apt_install "${PACKAGES_TO_INSTALL[@]}" Instead of (which I fixed):    apt_install ${PACKAGES_TO_INSTALL[@]}
Bash expansion of the second correct case puts each package as a separate argument. Putting them in double quotes puts a list of packages all as one
parameter, which isn't the name of a package, so it fails.

Also added case for arm64, where we need to actually build node-libcurl, which has some requirements. I put those directly in the install ff_agent
function.