Removed some commented out lines into the build tools project. i.e. things that only needed to be installed if we are building things. - Java certificates - Mono certificates
Also noticed that a function that installed things globally (and shouldn't) was:
a) defined twice!

So I removed it. The function was:
install_required_npm_libraries

It used to install as follows:

  # npm install --global commander  || { set_state "${FUNCNAME[0]}" "error_npm_install_commander"; return 1; }
  # npm install --global casperjs   || { set_state "${FUNCNAME[0]}" "error_npm_install_casperjs"; return 1; }
  # npm install --global json   || { set_state "${FUNCNAME[0]}" "error_npm_install_jsontool"; return 1; }
  # npm install --global slimerjs   || { set_state "${FUNCNAME[0]}" "error_npm_install_slimerjs"; return 1; }
  # npm install --global typescript || { set_state "${FUNCNAME[0]}" "error_npm_install_typescript"; return 1; }

Which makes no sense. We do not require commander, casperjs, json, slimerjs and typescript as part of our basic set_environment.
Typescript is for develoeprs anyway. Yikes!